+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Tools no longer take 2 damage when hurting entities
+
 ## 3.4.2
 * Coal veins are now rarer
 * Better Loot data pack now disables itself if ITE is present